Added character movement
The character movement isn't quite correct as the instruction and current version do not line up, but this is the partial progress for where I am at now.  Will be changing it with the camera as the 3rd person behind player view doesn't feel as good as I was hoping.